requirements.txt
click~=8.0.1
requests~=2.26.0
gitpython~=3.1.18
numexpr~=2.7.3